+++ NEWS.md
@@ -3,6 +3,7 @@
 * Improve mixed-geometry export error message (#78)
 * Remove NULL the_geom features from topojson output (#80)
 * Fix crash when issuing SQL "COPY" command
+* Return an error when "the_geom" is in skipfield for SVG output (#73)
 
 1.3.3 (11/01/13)
 -----